// apps/nest/src/enums/errors.enums.ts
export enum ErrorsMessagesEnum {
  NOT_FOUND = 'NOT_FOUND',
  USER_NOT_FOUND = "USER_NOT_FOUND",
  BAD_REQUEST = 'BAD_REQUEST',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  DATABASE_ERROR = 'DATABASE_ERROR',
  DUPLICATE_EMAIL = ' DUPLICATE_EMAIL',
  DUPLICATE_BOOK = 'DUPLICATE_BOOK',
  INVALID_PASSPORT = 'INVALID_PASSPORT',
  MISSING_REQUIRED_FIELD = 'MISSING_REQUIRED_FIELD',
  DATA_TOO_LONG = 'DATA_TOO_LONG',
  FOREIGN_KEY_CONSTRAINT = 'FOREIGN_KEY_CONSTRAINT',
  INVALID_DATA_TYPE = 'INVALID_DATA_TYPE',
}